+++ source/custom-resource/index.ts
@@ -1,16 +1,29 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import CloudFormation from "aws-sdk/clients/cloudformation";
-import EC2, { DescribeRegionsRequest } from "aws-sdk/clients/ec2";
-import ServiceCatalogAppRegistry from "aws-sdk/clients/servicecatalogappregistry";
-import S3, {
-  CreateBucketRequest,
-  PutBucketEncryptionRequest,
-  PutBucketPolicyRequest,
-  PutBucketVersioningRequest,
-} from "aws-sdk/clients/s3";
-import SecretsManager from "aws-sdk/clients/secretsmanager";
+import { CloudFormationClient, DescribeStackResourcesCommand } from "@aws-sdk/client-cloudformation";
+import { EC2Client, DescribeRegionsCommand, DescribeRegionsCommandInput } from "@aws-sdk/client-ec2";
+import { ServiceCatalogAppRegistryClient, GetApplicationCommand } from "@aws-sdk/client-service-catalog-appregistry";
+import {
+  S3Client,
+  HeadBucketCommand,
+  HeadObjectCommand,
+  PutObjectCommand,
+  CreateBucketCommand,
+  PutBucketEncryptionCommand,
+  PutBucketPolicyCommand,
+  PutBucketTaggingCommand,
+  PutBucketVersioningCommand,
+  GetBucketLocationCommand,
+  PutBucketPolicyCommandInput,
+  PutBucketTaggingCommandInput,
+  PutBucketEncryptionCommandInput,
+  PutBucketVersioningCommandInput,
+  CreateBucketCommandInput,
+  PutObjectCommandInput,
+} from "@aws-sdk/client-s3";
+import { SecretsManagerClient, GetSecretValueCommand } from "@aws-sdk/client-secrets-manager";
+import { CloudFrontClient, GetDistributionCommand } from "@aws-sdk/client-cloudfront";
 import axios, { RawAxiosRequestConfig, AxiosResponse } from "axios";
 import { createHash } from "crypto";
 import moment from "moment";
@@ -39,15 +52,13 @@ import {
   GetAppRegApplicationNameRequestProperties,
   ValidateExistingDistributionRequestProperties,
 } from "./lib";
-import CloudFront from "aws-sdk/clients/cloudfront";
-
 const awsSdkOptions = getOptions();
-const s3Client = new S3(awsSdkOptions);
-const ec2Client = new EC2(awsSdkOptions);
-const cloudformationClient = new CloudFormation(awsSdkOptions);
-const serviceCatalogClient = new ServiceCatalogAppRegistry(awsSdkOptions);
-const secretsManager = new SecretsManager(awsSdkOptions);
-const cloudfrontClient = new CloudFront(awsSdkOptions);
+const s3Client = new S3Client({ ...awsSdkOptions, followRegionRedirects: true });
+const ec2Client = new EC2Client(awsSdkOptions);
+const cloudformationClient = new CloudFormationClient(awsSdkOptions);
+const serviceCatalogClient = new ServiceCatalogAppRegistryClient(awsSdkOptions);
+const secretsManager = new SecretsManagerClient(awsSdkOptions);
+const cloudfrontClient = new CloudFrontClient(awsSdkOptions);
 
 const { SOLUTION_ID, SOLUTION_VERSION, AWS_REGION, RETRY_SECONDS } = process.env;
 const METRICS_ENDPOINT = "https://metrics.awssolutionsbuilder.com/generic";
@@ -359,7 +370,7 @@ async function putConfigFile(
   const content = `'use strict';\n\nconst appVariables = {\n${configFieldValues}\n};`;
 
   // In case getting object fails due to asynchronous IAM permission creation, it retries.
-  const params = {
+  const params: PutObjectCommandInput = {
     Bucket: DestS3Bucket,
     Body: content,
     Key: DestS3key,
@@ -370,7 +381,7 @@ async function putConfigFile(
     try {
       console.info(`Putting ${DestS3key}... Try count: ${retry}`);
 
-      await s3Client.putObject(params).promise();
+      await s3Client.send(new PutObjectCommand(params));
 
       console.info(`Putting ${DestS3key} completed.`);
       break;
@@ -417,9 +428,8 @@ async function validateBuckets(requestProperties: CheckSourceBucketsRequestPrope
   const errorBuckets = [];
 
   for (const bucket of checkBuckets) {
-    const params = { Bucket: bucket };
     try {
-      await s3Client.headBucket(params).promise();
+      await s3Client.send(new HeadBucketCommand({ Bucket: bucket }));
 
       console.info(`Found bucket: ${bucket}`);
     } catch (error) {
@@ -471,25 +481,24 @@ async function checkFirstBucketRegion(
   console.info(`Attempting to check if the following bucket exists in the same region as deployment: ${bucket}`);
 
   try {
-    const bucketLocation = await s3Client.getBucketLocation({ Bucket: bucket }).promise();
-    const bucketRegion = bucketLocation.LocationConstraint || "us-east-1";
+    const bucketLocationResponse = await s3Client.send(new GetBucketLocationCommand({ Bucket: bucket }));
+    const bucketRegion = bucketLocationResponse.LocationConstraint || "us-east-1";
     if (bucketRegion === AWS_REGION) {
       console.info(`Bucket '${bucket}' is in the same region (${bucketRegion}) as the S3 client.`);
       return { BucketName: bucket, BucketHash: generateBucketHash(bucket) };
     } else {
       try {
-        const params = { Bucket: dummyBucketName };
-        await s3Client.headBucket(params).promise();
+        await s3Client.send(new HeadBucketCommand({ Bucket: dummyBucketName }));
 
         console.info(`Found bucket: ${dummyBucketName}`);
         return { BucketName: dummyBucketName, BucketHash: generateBucketHash(dummyBucketName) };
       } catch (error) {
         console.info(`Could not find dummy bucket. Creating bucket in region: ${AWS_REGION}`);
-        await s3Client.createBucket({ Bucket: dummyBucketName }).promise();
+        await s3Client.send(new CreateBucketCommand({ Bucket: dummyBucketName }));
         try {
           console.info("Adding tag...");
 
-          const taggingParams = {
+          const taggingParams: PutBucketTaggingCommandInput = {
             Bucket: dummyBucketName,
             Tagging: {
               TagSet: [
@@ -500,7 +509,7 @@ async function checkFirstBucketRegion(
               ],
             },
           };
-          await s3Client.putBucketTagging(taggingParams).promise();
+          await s3Client.send(new PutBucketTaggingCommand(taggingParams));
 
           console.info(`Successfully added tag to bucket '${dummyBucketName}'`);
         } catch (error) {
@@ -526,18 +535,18 @@ async function getAppRegApplicationName(
   requestProperties: GetAppRegApplicationNameRequestProperties
 ): Promise<{ ApplicationName?: string }> {
   try {
-    const stackResources = await cloudformationClient
-      .describeStackResources({
+    const stackResources = await cloudformationClient.send(
+      new DescribeStackResourcesCommand({
         StackName: requestProperties.StackId,
         LogicalResourceId: "AppRegistry968496A3",
       })
-      .promise();
+    );
 
-    const application = await serviceCatalogClient
-      .getApplication({
+    const application = await serviceCatalogClient.send(
+      new GetApplicationCommand({
         application: stackResources.StackResources[0].PhysicalResourceId,
       })
-      .promise();
+    );
     return {
       ApplicationName: application?.name ?? requestProperties.DefaultName,
     };
@@ -558,11 +567,11 @@ async function validateExistingDistribution(
   requestProperties: ValidateExistingDistributionRequestProperties
 ): Promise<{ DistributionDomainName?: string }> {
   try {
-    const response = await cloudfrontClient
-      .getDistribution({
+    const response = await cloudfrontClient.send(
+      new GetDistributionCommand({
         Id: requestProperties.ExistingDistributionID,
       })
-      .promise();
+    );
 
     return { DistributionDomainName: response.Distribution?.DomainName };
   } catch (error) {
@@ -593,7 +602,7 @@ async function checkSecretsManager(
 
   for (let retry = 1; retry <= RETRY_COUNT; retry++) {
     try {
-      const response = await secretsManager.getSecretValue({ SecretId: SecretsManagerName }).promise();
+      const response = await secretsManager.send(new GetSecretValueCommand({ SecretId: SecretsManagerName }));
       const secretString = JSON.parse(response.SecretString);
 
       if (!Object.prototype.hasOwnProperty.call(secretString, SecretsManagerKey)) {
@@ -648,7 +657,7 @@ async function checkFallbackImage(
 
   for (let retry = 1; retry <= RETRY_COUNT; retry++) {
     try {
-      data = await s3Client.headObject({ Bucket: FallbackImageS3Bucket, Key: FallbackImageS3Key }).promise();
+      data = await s3Client.send(new HeadObjectCommand({ Bucket: FallbackImageS3Bucket, Key: FallbackImageS3Key }));
       break;
     } catch (error) {
       if (retry === RETRY_COUNT || ![ErrorCodes.ACCESS_DENIED, ErrorCodes.FORBIDDEN].includes(error.code)) {
@@ -693,28 +702,27 @@ async function createCloudFrontLoggingBucket(requestProperties: CreateLoggingBuc
     `The opt-in status of the '${AWS_REGION}' region is '${isOptInRegion ? "opted-in" : "opt-in-not-required"}'`
   );
 
+  const regionS3Client = new S3Client({
+    ...awsSdkOptions,
+    region: targetRegion,
+  });
+
   // create bucket
   try {
-    const s3Client = new S3({
-      ...awsSdkOptions,
-      apiVersion: "2006-03-01",
-      region: targetRegion,
-    });
-
-    const createBucketRequestParams: CreateBucketRequest = {
+    const createBucketRequestParams: CreateBucketCommandInput = {
       Bucket: bucketName,
-      ACL: "log-delivery-write",
-      ObjectOwnership: "ObjectWriter",
+      ACL: "log-delivery-write" as never, // Type assertion to suppress ACL and ownership, no type-def available for the needed values
+      ObjectOwnership: "ObjectWriter" as never,
     };
-    await s3Client.createBucket(createBucketRequestParams).promise();
+    await regionS3Client.send(new CreateBucketCommand(createBucketRequestParams));
 
     console.info(`Successfully created bucket '${bucketName}' in '${targetRegion}' region`);
 
-    const putBucketVersioningRequestParams: PutBucketVersioningRequest = {
+    const putBucketVersioningRequestParams: PutBucketVersioningCommandInput = {
       Bucket: bucketName,
       VersioningConfiguration: { Status: "Enabled" },
     };
-    await s3Client.putBucketVersioning(putBucketVersioningRequestParams).promise();
+    await regionS3Client.send(new PutBucketVersioningCommand(putBucketVersioningRequestParams));
     console.info(`Successfully enabled versioning on '${bucketName}'`);
   } catch (error) {
     console.error(`Could not create bucket '${bucketName}' or failed to enable versioning`);
@@ -726,14 +734,14 @@ async function createCloudFrontLoggingBucket(requestProperties: CreateLoggingBuc
   // add encryption to bucket
   console.info("Adding Encryption...");
   try {
-    const putBucketEncryptionRequestParams: PutBucketEncryptionRequest = {
+    const putBucketEncryptionRequestParams: PutBucketEncryptionCommandInput = {
       Bucket: bucketName,
       ServerSideEncryptionConfiguration: {
         Rules: [{ ApplyServerSideEncryptionByDefault: { SSEAlgorithm: "AES256" } }],
       },
     };
 
-    await s3Client.putBucketEncryption(putBucketEncryptionRequestParams).promise();
+    await regionS3Client.send(new PutBucketEncryptionCommand(putBucketEncryptionRequestParams));
 
     console.info(`Successfully enabled encryption on bucket '${bucketName}'`);
   } catch (error) {
@@ -746,7 +754,6 @@ async function createCloudFrontLoggingBucket(requestProperties: CreateLoggingBuc
   // add policy to bucket
   try {
     console.info("Adding policy...");
-
     const bucketPolicyStatement = {
       Resource: `arn:aws:s3:::${bucketName}/*`,
       Action: "*",
@@ -759,12 +766,12 @@ async function createCloudFrontLoggingBucket(requestProperties: CreateLoggingBuc
       Version: "2012-10-17",
       Statement: [bucketPolicyStatement],
     };
-    const putBucketPolicyRequestParams: PutBucketPolicyRequest = {
+    const putBucketPolicyRequestParams: PutBucketPolicyCommandInput = {
       Bucket: bucketName,
       Policy: JSON.stringify(bucketPolicy),
     };
 
-    await s3Client.putBucketPolicy(putBucketPolicyRequestParams).promise();
+    await regionS3Client.send(new PutBucketPolicyCommand(putBucketPolicyRequestParams));
 
     console.info(`Successfully added policy to bucket '${bucketName}'`);
   } catch (error) {
@@ -777,8 +784,7 @@ async function createCloudFrontLoggingBucket(requestProperties: CreateLoggingBuc
   // Add Stack tag
   try {
     console.info("Adding tag...");
-
-    const taggingParams = {
+    const taggingParams: PutBucketTaggingCommandInput = {
       Bucket: bucketName,
       Tagging: {
         TagSet: [
@@ -789,7 +795,7 @@ async function createCloudFrontLoggingBucket(requestProperties: CreateLoggingBuc
         ],
       },
     };
-    await s3Client.putBucketTagging(taggingParams).promise();
+    await regionS3Client.send(new PutBucketTaggingCommand(taggingParams));
 
     console.info(`Successfully added tag to bucket '${bucketName}'`);
   } catch (error) {
@@ -807,11 +813,11 @@ async function createCloudFrontLoggingBucket(requestProperties: CreateLoggingBuc
  * @returns The result of check.
  */
 async function checkRegionOptInStatus(region: string): Promise<boolean> {
-  const describeRegionsRequestParams: DescribeRegionsRequest = {
+  const describeRegionsRequestParams: DescribeRegionsCommandInput = {
     RegionNames: [region],
     Filters: [{ Name: "opt-in-status", Values: ["opted-in"] }],
   };
-  const describeRegionsResponse = await ec2Client.describeRegions(describeRegionsRequestParams).promise();
+  const describeRegionsResponse = await ec2Client.send(new DescribeRegionsCommand(describeRegionsRequestParams));
 
   return describeRegionsResponse.Regions.length > 0;
 }